WEB: Improve inheritance for Section classes
