+++ include/Objects/DownloadsSection.php
@@ -6,29 +6,64 @@ namespace ScummVM\Objects;
  */
 class DownloadsSection extends BasicSection
 {
+    private $notes;
+    private $footer;
+    private $files;
+    private $links;
+    private $items;
 
-    /* DownloadsSection object constructor. */
     public function __construct($data)
     {
         parent::__construct($data);
-        $this->subsections = array();
+        $this->notes = $data['notes'];
+        $this->footer = $data['footer'];
+        $this->files = array();
+        $this->links = array();
+        $this->items = array();
 
-        parent::toArray($data['subsection']);
-        foreach ($data['subsection'] as $key => $value) {
-            $this->subsections[] = new DSubSection($value);
+        if (isset($data['entries'])) {
+            foreach ($data['entries'] as $type => $item) {
+                parent::toArray($item);
+                if ($type == 'file') {
+                    foreach ($item as $file) {
+                        $this->items[] = new File($file);
+                    }
+                } elseif ($type == 'link') {
+                    foreach ($item as $link) {
+                        $this->items[] = new WebLink($link);
+                    }
+                }
+            }
         }
     }
 
-
-    /* Get the base URL. */
-    public function getBaseURL()
+    /* Get the optional notes. */
+    public function getNotes()
     {
-        return $this->baseurl;
+        return $this->notes;
     }
 
-    /* Get the list of optional subsections. */
-    public function getSubSections()
+    /* Get the optional footer. */
+    public function getFooter()
     {
-        return $this->subsections;
+        return $this->footer;
+    }
+
+    /* Get the list of files. */
+    public function getFiles()
+    {
+        return $this->files;
+    }
+
+    /* Get the list of links. */
+    public function getLinks()
+    {
+        return $this->links;
+    }
+
+    /* Get the list of items. */
+    public function getItems()
+    {
+        return $this->items;
     }
 }
